code-revision: moving the animations to "clean" exported calls
